move next_proof to own module
